requirements.txt
customtkinter
requests
PyJWT
urllib3
zeroconf>=0.131.0
